pyproject.toml
[tool.poetry]
name = "online-academy"
version = "0.1.0"
description = ""
authors = ["exetch <gassemark@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "online_academy"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
pillow = "9.5.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
faker = "^20.0.3"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
